implement GetMinTS on the API leader
